move utils.test.tsx
