+++ containers/ecr-viewer/src/app/view-data/utils/utils.tsx
